# sparse model fields