In the same spirit as #15370

https://code.djangoproject.com/ticket/33502#ticket

(I'm not actually sure it actually makes much of a difference in practice.)