@kezabelle 2022-02-08
So ignoring the unverified runtime cost/benefit (see ticket, + #15126), and accepting the fact that it's early and I'm patching in ... at-best a rough approximation, here's my findings, running across the test suite (using `--parallel=1`, `14987` tests ran) using a global dict (ick) and a `pympler.asizeof.Asizer()` instance to hopefully account for duplicates/double-counting, results were measured in bytes per `type(self)` where `self` is a `Field` at the end of `__init__`, and output using the `filesizeformat` filter. Python version is `3.10.1` in case the internal representation sizes have changed.

Baseline, as of 25514b604a64686ba603bf10a8a63390dc38b79d. I've snipped everything that's under 1MB for brevity's sake:
```
<class 'django.db.models.fields.files.FileField'> accounted for 1.5 MB
<class 'django.db.models.fields.Field'> accounted for 1.7 MB
<class 'django.db.models.fields.TextField'> accounted for 2.2 MB
<class 'django.db.models.fields.PositiveIntegerField'> accounted for 2.3 MB
<class 'django.db.models.fields.SlugField'> accounted for 2.3 MB
<class 'django.db.models.fields.DateField'> accounted for 2.6 MB
<class 'django.db.models.fields.DateTimeField'> accounted for 3.4 MB
<class 'django.db.models.fields.BooleanField'> accounted for 3.4 MB
<class 'django.db.models.fields.related.OneToOneField'> accounted for 5.6 MB
<class 'django.db.models.fields.related.ManyToManyField'> accounted for 7.2 MB
<class 'django.db.models.fields.IntegerField'> accounted for 14.2 MB
<class 'django.db.models.fields.related.ForeignKey'> accounted for 16.6 MB
<class 'django.db.models.fields.CharField'> accounted for 44.8 MB
<class 'django.db.models.fields.AutoField'> accounted for 57.9 MB
```
And then after manually applying this patch set, same ordering:
```
<class 'django.db.models.fields.files.FileField'> accounted for 722.7 KB
<class 'django.db.models.fields.Field'> accounted for 1.3 MB
<class 'django.db.models.fields.TextField'> accounted for 1.8 MB
<class 'django.db.models.fields.PositiveIntegerField'> accounted for 1.9 MB
<class 'django.db.models.fields.SlugField'> accounted for 1.1 MB
<class 'django.db.models.fields.DateField'> accounted for 2.4 MB
<class 'django.db.models.fields.DateTimeField'> accounted for 3.2 MB
<class 'django.db.models.fields.BooleanField'> accounted for 3.2 MB
<class 'django.db.models.fields.related.OneToOneField'> accounted for 6.8 MB
<class 'django.db.models.fields.related.ManyToManyField'> accounted for 5.7 MB
<class 'django.db.models.fields.IntegerField'> accounted for 9.3 MB
<class 'django.db.models.fields.related.ForeignKey'> accounted for 20.3 MB
<class 'django.db.models.fields.CharField'> accounted for 26.8 MB
<class 'django.db.models.fields.AutoField'> accounted for 39.2 MB
```

It probably would've made sense for me to output the **total** of bytes used by `Field` classes ... but I didn't think to, and it'd be another `1000secs` of my battery's lifetime to do so now ;)

Eagle eyed viewers may spot that generally the numbers go down (good!) but there's possibly some discrepancy that I've not accounted for, because some of of the `related` fields end up using _more_. Perhaps that's a measurement error on my part, or perhaps it's to do with their custom descriptors, or the values which get pushed into them being tracked by some nuance of `pympler`.

A quick peek at the change on a per-`Field` basis:
```
# without class attributes, without anything being set to something other than `None`
In [3]: asizeof(Field())
Out[3]: 4584
# with class attributes, no parameters being sent, so everything is again `None`
In [3]: asizeof(Field())
Out[3]: 3040
```