@collinanderson 2022-02-08
@kezabelle could you share your `pympler` testing code?

I rebased and optimized AutoField and related fields a bit if you're interested here:
https://github.com/django/django/compare/main...collinanderson:sparsefields

Yes, I realize attribute access is theoretically slightly slower, but a lot of these attributes are only accessed for validation and migrations, and the performance hit really shouldn't be much. Also, there are already other attributes defined on the model.

Some test-suite run times, seems to not affect speed much, but hopefully at least reduces memory usage a little.

python3.10 MacOS M1:  time ./runtests.py (no parallel, skipped=1666)
without patch:
150.61s user, 10.49s system, 64% cpu, 4:09.78 total
150.22s user, 10.88s system, 64% cpu, 4:09.93 total
150.12s user 10.45s system 64% cpu 4:09.30 total
151.05s user 10.68s system 64% cpu 4:10.52 total
150.69s user 10.69s system 64% cpu 4:10.08 total
with patch:
150.34s user, 10.88s system, 64% cpu, 4:10.09 total
149.46s user, 10.44s system, 64% cpu, 4:08.88 total
149.56s user, 10.58s system, 64% cpu, 4:08.85 total
149.52s user 10.62s system 64% cpu 4:08.98 total
150.08s user 10.69s system 64% cpu 4:09.40 total
150.44s user 10.63s system 64% cpu 4:09.81 total

python3.8 linux 4-core shared vps: time ./runtests.py --parallel=4 (skipped=1216)
without patch:
real    3m34.918s, user    9m22.476s, sys     0m54.566s
real    3m33.835s, user    9m20.487s, sys     0m48.310s
real    3m33.985s, user    9m24.748s, sys     0m51.688s
real    3m34.405s, user    9m18.959s, sys     0m53.592s
real    3m33.877s, user    9m15.704s, sys     0m53.087s
real    3m35.571s, user    9m23.005s, sys     0m53.644s
with patch:
real    3m34.070s, user    9m16.471s, sys     0m52.904s
real    3m32.681s, user    9m16.295s, sys     0m51.821s
real    3m31.244s, user    9m10.681s, sys     0m52.282s
real    3m34.017s, user    9m17.179s, sys     0m45.538s
real    3m31.971s, user    9m15.597s, sys     0m51.727s
real    3m33.401s, user    9m18.072s, sys     0m52.998s
real    3m35.312s, user    9m20.305s, sys     0m54.061s