@kezabelle 2022-02-08
> Could you share your pympler testing code?

Sure. It's lazy and  quick and dirty (I _did_ say it was early ;)), and there's absolutely scope for me to be miscounting, so I've only been taking it as a guide (if a `Field` is `3040b` it stands to reason that they're all at least that, give or take de-duplication, regardless of what the overall summary is). It's _probably_ "OK" enough to get a flavour pre/post change, even if it's not fully accurate and correct. Or so I tell myself. Please point out any improvements or accountable errors you can think of so I can do them better next time...

Pushed the commit as aeb47907c973d42a09cfb295899070e7fdbc5999, on top of which I applied your changes manually. Pretty sure to even hope to be usefully correct, you'd have to be using `--parallel=1` as I have been.

> Yes, I realize attribute access is theoretically slightly slower, but a lot of these attributes are only accessed for validation and migrations, and the performance hit really shouldn't be much.

Yeah, the difference I saw on the other PR was minimal, and at least for _setting_ the values, only looked to affect things by a couple of `ns` _once_ (after which it's in the instance dict). I'm still hopeful _that_ PR ends up being objectively and demonstrably OK for consideration, though it needs more eyes on it to check my theory. In _theory_ I believe the same would hold true here (given the way Fields specifically get used and held onto), but I'm inherently biased having gone through a similar idea, so I can't in good faith assist in being a deciding factor - I can only help try and gather anecdotal data and see if it matches our expectations and hopes.